+++ src/Command/InfectionCommand.php
@@ -27,9 +27,9 @@ use Infection\Process\Runner\MutationTestingRunner;
 use Infection\Process\Runner\TestRunConstraintChecker;
 use Infection\TestFramework\Coverage\CodeCoverageData;
 use Infection\TestFramework\MemoryUsageAware;
-use Infection\TestFramework\PhpSpec\ExtraOptions;
+use Infection\TestFramework\PhpSpec\ExtraOptions as PhpSpecExtraOptions;
 use Infection\TestFramework\PhpUnit\Coverage\CoverageXmlParser;
-use Infection\TestFramework\PhpUnit\ExtraOptions;
+use Infection\TestFramework\PhpUnit\ExtraOptions as PhpUnitExtraOptions;
 use Infection\TestFramework\TestFrameworkExtraOptions;
 use Infection\TestFramework\TestFrameworkTypes;
 use Symfony\Component\Console\Input\ArrayInput;
@@ -341,8 +341,8 @@ final class InfectionCommand extends BaseCommand
         $extraOptions = $this->input->getOption('test-framework-options');
 
         return TestFrameworkTypes::PHPUNIT === $testFrameworkKey
-            ? new ExtraOptions($extraOptions)
-            : new ExtraOptions($extraOptions);
+            ? new PhpUnitExtraOptions($extraOptions)
+            : new PhpSpecExtraOptions($extraOptions);
     }
 
     /**
