+++ tests/TestFramework/PhpSpec/AdapterTest.php
@@ -7,7 +7,7 @@
 
 declare(strict_types=1);
 
-namespace Infection\Tests\TestFramework\PhpSpec\Adapter;
+namespace Infection\Tests\TestFramework\PhpSpec;
 
 use Infection\Finder\AbstractExecutableFinder;
 use Infection\TestFramework\CommandLineArgumentsAndOptionsBuilder;
@@ -20,7 +20,7 @@ use Mockery;
 /**
  * @internal
  */
-final class PhpSpecAdapterTest extends Mockery\Adapter\Phpunit\MockeryTestCase
+final class AdapterTest extends Mockery\Adapter\Phpunit\MockeryTestCase
 {
     public function test_it_has_a_name(): void
     {
